Modified files in Frontend.
