+++ frontend/src/App.jsx
@@ -4,12 +4,13 @@ import HomePage from "./components/HomePage";
 import { useAuth } from "./context/AuthContext";
 import RecipePage from "./components/recipePage/RecipePage";
 import { Routes, Route } from "react-router-dom";
-import ShoppingList from "./components/ShoppingList";
-import ShoppingListReview from "./components/ShoppingListReview";
 import PhotoAlbum from "./components/PhotoAlbum";
-import AddIngredient from "./components/AddIngredient";
-import AddShoppingListModal from "./components/shoppingListPage/AddShoppingListModal";
-import AllIngredients from "./components/AllIngredients";
+import AddIngredient from "./components/ingredientPage/AddIngredient";
+import AllIngredients from "./components/ingredientPage/AllIngredients";
+import CreateShoppingLists from "./components/shoppingListPage/CreateShoppingLists";
+import ShoppingList from "./components/shoppingListPage/ShoppingList";
+import ShoppingListReview from "./components/shoppingListPage/ShoppingListReview";
+
 
 function App() {
   const { token } = useAuth();
@@ -30,7 +31,7 @@ function App() {
           <Route path="/photoalbum" element={<PhotoAlbum />} />
           <Route path="/add-ingredient" element={<AddIngredient userId={1}/>} />
           <Route path="/all-ingredients" element={<AllIngredients userId={1}/>} />
-          <Route path="/create-shoppinglists" element={<AddShoppingListModal userId={1}/>} />
+          <Route path="/create-shoppinglists" element={<CreateShoppingLists userId={1}/>} />
         </Routes>
       </div>
     </main>
